List images for partial functions.
To define partial functions I now use a list of appropriate size (given by a dependent type family) as the range of functions rather than a list of pair. Makes it possible to reuse the notion of finite set more directly. The proofs are significantly simpler, and can be split into smaller units of meaning.

This follows a suggestion by Cyril Cohen.